[FIX] Move AuthError and TokenError Exception files to exception folder - 토큰 및 인증 관련 예외 처리 파일 이동
related to #1
