+++ src/main/java/date_naeun/naeunserver/api/SocialLoginApiController.java
@@ -3,9 +3,9 @@ package date_naeun.naeunserver.api;
 import date_naeun.naeunserver.config.jwt.JwtProvider;
 import date_naeun.naeunserver.config.jwt.dto.TokenDto;
 import date_naeun.naeunserver.config.jwt.dto.TokenRefreshDto;
-import date_naeun.naeunserver.config.exception.AuthErrorException;
-import date_naeun.naeunserver.config.exception.TokenStatus;
-import date_naeun.naeunserver.config.exception.TokenErrorException;
+import date_naeun.naeunserver.exception.AuthErrorException;
+import date_naeun.naeunserver.exception.TokenStatus;
+import date_naeun.naeunserver.exception.TokenErrorException;
 import date_naeun.naeunserver.dto.ResultDto;
 import date_naeun.naeunserver.exception.HttpStatusCode;
 import date_naeun.naeunserver.external.client.kakao.dto.KakaoUserInfo;
